requirements.txt
asgiref==3.4.1
Django==3.2.10
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
geographiclib==1.52
geopy==2.2.0
mysqlclient==2.1.0
PyJWT==2.3.0
PyMySQL==1.0.2
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2